+++ assets/javascripts/kitten/components/meters/progress.js
@@ -7,6 +7,7 @@ export class Progress extends Component {
       className,
       color,
       value,
+      rampProps,
       ...others,
     } = this.props
 
@@ -20,6 +21,11 @@ export class Progress extends Component {
       width: `${value}%`,
     }
 
+    const rampClassName = classNames(
+      'k-Progress__ramp',
+      rampProps.className,
+    )
+
     return (
       <div
         { ...others }
@@ -28,7 +34,7 @@ export class Progress extends Component {
         aria-valuemax="100"
         aria-valuenow={ this.props.value }
         className={ progressClassName }>
-        <div className="k-Progress__ramp">
+        <div { ...rampProps } className={ rampClassName }>
           <div
             className="k-Progress__slider"
             style={ style }
@@ -43,4 +49,5 @@ Progress.defaultProps = {
   className: null,
   color: null,
   value: 50,
+  rampProps: {},
 }