+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Breaking change: Rename `style` prop to `fontStyle` in `Text` component.
+- Feature: Add `rampProps` prop to `Progress` component.
+- Feature: Add `CrowdfundingCard` component.
 - Feature: Add `mediaQueries` HOC.
 - Fix: Update `primary-light` color.
 - Feature: Add `modifier` prop on `TagButton` component.